move handleProjectDownload to projectDownload.tsx
