--- packages/ui/src/components/editor/panels/Files/download/projectDownloadProgress.tsx
@@ -1,45 +1,0 @@
-/*
-CPAL-1.0 License
-
-The contents of this file are subject to the Common Public Attribution License
-Version 1.0. (the "License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-https://github.com/EtherealEngine/etherealengine/blob/dev/LICENSE.
-The License is based on the Mozilla Public License Version 1.1, but Sections 14
-and 15 have been added to cover use of software over a computer network and 
-provide for limited attribution for the Original Developer. In addition, 
-Exhibit A has been modified to be consistent with Exhibit B.
-
-Software distributed under the License is distributed on an "AS IS" basis,
-WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
-specific language governing rights and limitations under the License.
-
-The Original Code is Ethereal Engine.
-
-The Original Developer is the Initial Developer. The Initial Developer of the
-Original Code is the Ethereal Engine team.
-
-All portions of the code written by the Ethereal Engine team are Copyright © 2021-2023 
-Ethereal Engine. All Rights Reserved.
-*/
-
-import React from 'react'
-import { useTranslation } from 'react-i18next'
-import Progress from '../../../../../primitives/tailwind/Progress'
-
-export const ProjectDownloadProgress = ({ completed, total, progress, isDownloading }) => {
-  const { t } = useTranslation()
-
-  return isDownloading ? (
-    <div className="flex h-auto w-full justify-center pb-2 pt-2">
-      <div className="flex w-1/2">
-        <span className="inline-block pr-2 text-xs font-normal leading-none text-theme-primary">
-          {t('editor:layout.filebrowser.downloadingProject', { completed, total })}
-        </span>
-        <div className="basis-1/2">
-          <Progress value={progress} />
-        </div>
-      </div>
-    </div>
-  ) : null
-}
